consolide screenshot code back into scaler.cpp, remove custom resize/bmp code.
