+++ scaler.h
@@ -11,6 +11,7 @@ typedef enum {
     BGR,
     BGRA,
     RGBA,
+    ARGB32, // respect endianness
     YUV,
 } mister_scaler_format_t;
 
@@ -31,10 +32,10 @@ typedef struct {
 #define MISTER_SCALER_BUFFERSIZE   2048*3*1024
 
 mister_scaler *mister_scaler_init();
-int mister_scaler_read(mister_scaler *,unsigned char *buffer, mister_scaler_format_t format);
-//int mister_scaler_read_bgr(mister_scaler *ms, unsigned char *gbuf, mister_scaler_format_t format);
-//int mister_scaler_read_32(mister_scaler *ms, unsigned char *buffer, mister_scaler_format_t format);
-//int mister_scaler_read_yuv(mister_scaler *ms,int,unsigned char *y,int, unsigned char *U,int, unsigned char *V);
+int mister_scaler_read(mister_scaler *,unsigned char *buffer, mister_scaler_format_t format = ARGB32);
 void mister_scaler_free(mister_scaler *);
+
+void request_screenshot(char *cmd, int scaled = 0);
+void screenshot_cb(void);
 
 #endif
